dev/backend/requirements.txt
bcrypt==3.1.6
Cerberus==1.3.1
cffi==1.12.2
Click==7.0
coverage==4.5.3
dnspython
ecdsa==0.13.2
Eve==0.9.2
eve-healthcheck==0.2.0
Events==0.3
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.6
Flask-PyMongo==0.5.1
future==0.17.1
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10.1
MarkupSafe==0.23
py-healthcheck==1.8.0
pyasn1==0.4.5
pycparser==2.19
pymongo==3.7.2
python-dotenv==0.6.5
python-jose==3.0.1
redis==3.3.7
rsa==4.0
simplejson==3.16.0
six==1.12.0
Werkzeug==0.15.4
